🚧 scx: properly disable scx schedulers
the way we were disabling scx schedulers was completely invalid in the
current setup (after introduction of scx_loader), so for us to correctly
disable it we need to do these steps:
1. set `default_sched` to None.
2. edit config file.
3. call method to stop current scheduler.

we cannot disable scx_loader, because the KernelManager is using it to
get information. so it makes no sense in stoping the loader it self.